Changed code, new code added
- MonthlyCostsResults entity added,
- MonthlyCostsResultsRepository added,
- MonthlyCosts entity modified,
- MonthlyCostsService modified,
- UserResource RestController method sumUpAllTheMonthlyCosts modified

+++ src/main/java/com/jerzymaj/budgetmanagement/budget_management_app/jpa_repositories/MonthlyCostsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MonthlyCostsRepository extends JpaRepository<MonthlyCosts, Long> {
+public interface MonthlyCostsRepository extends JpaRepository<MonthlyCosts, Integer> {
 
     Optional<MonthlyCosts> findByUserId(Long userId);
